P2Plending-Prediction-main/requirements.txt
absl-py==1.0.0
altair==4.1.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.0
bleach==4.1.0
blinker==1.4
bokeh==2.4.3
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
colorcet==3.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==4.0
entrypoints==0.3
exceptiongroup==1.1.3
fastjsonschema==2.15.3
Flask==2.0.2
Flask-WTF==1.0.1
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
glfw==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.14.0
h5py==3.6.0
holoviews==1.15.2
hvplot==0.8.2
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
json5==0.9.6
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.16.0
jupyterlab==3.3.3
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multipledispatch==0.6.0
mysql-connector-python==8.0.30
nbclassic==0.3.7
nbclient==0.5.9
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.4
nltk==3.6.6
notebook==6.4.6
notebook-shim==0.1.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
panel==0.14.1
param==1.12.2
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.14.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyct==0.4.8
pydeck==0.7.1
Pygments==2.10.0
Pympler==1.0.1
pyngrok==5.1.0
PyOpenGL==3.1.5
pyparsing==3.0.6
pyrr==0.10.3
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.2.2
scipy==1.7.3
seaborn==0.12.1
selenium==4.12.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
sqlparse==0.4.2
streamlit==1.3.0
tenacity==8.2.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.3
zipp==3.6.0
